implemented generic search and count handler
